data-consumer/src/requirements.txt
requests
kafka-python
